Fix floating point errors

+++ RobotControl.java
@@ -13,7 +13,7 @@ import icommand.nxt.comm.NXTCommand;
 public class RobotControl {
 
 	// Default ratio between wheels when turning (e.g. with a factor of 2, outer wheel is 2x faster than pivot wheel)
-	private static final double DEFAULT_SPEED_FACTOR = 2.0;
+	private static final int DEFAULT_SPEED_FACTOR = 2;
 
 	// Default frequency used for beeping
 	private static final int DEFAULT_BEEP_FREQ = 500;
@@ -203,10 +203,10 @@ public class RobotControl {
 
 	/**
 	 * Instruct robot to turn left
-	 * @param speedFactor double Ratio between the speeds of the two wheels: higher the ratio, the sharper the turn
+	 * @param speedFactor int Ratio between the speeds of the two wheels: higher the ratio, the sharper the turn
 	 * @return Delayer Used to chain waitFor() to provide a minimum duration of movement is achieved
 	 */
-	public static Delayer goLeft(double speedFactor) {
+	public static Delayer goLeft(int speedFactor) {
 		// Change zero and negative values to a default, preventing unwanted / dangerous behaviour		
 		if(speedFactor > 0) {
 			MOTOR_LEFT.setSpeed(baseSpeed / speedFactor);
@@ -230,10 +230,10 @@ public class RobotControl {
 
 	/**
 	 * Instruct robot to turn right
-	 * @param speedFactor double Ratio between the speeds of the two wheels: higher the ratio, the sharper the turn
+	 * @param speedFactor int Ratio between the speeds of the two wheels: higher the ratio, the sharper the turn
 	 * @return Delayer Used to chain waitFor() to provide a minimum duration of movement is achieved
 	 */
-	public static Delayer goRight(double speedFactor) {
+	public static Delayer goRight(int speedFactor) {
 		// Change zero and negative values to a default, preventing unwanted / dangerous behaviour
 		if(speedFactor > 0) {
 			MOTOR_LEFT.setSpeed(baseSpeed);
